팀원: 김재섭
메인, 관리자 리뷰 리스트 작성중

+++ Babiyo/src/main/java/com/mealmaker/babiyo/review/model/ReviewDto.java
@@ -12,7 +12,7 @@ public class ReviewDto {
 	private String productName;
 	private String categoryName;
 	private String memberId;
-	private int starRating;
+	private double starRating;
 	private String content;
 	private Date createDate;
 	private Date modifyDate;
